// modules/script-module-metadata.d.ts
export interface MetadataDefinition {
   name: string;
}
export interface MetadataTypedDefinition {
   type: MetadataType;
}
export interface MetadataMemberDefinition extends MetadataDefinition {
   is_static: boolean;
   is_read_only: boolean;
}
// ...
export type MetadataType = {
   is_bind_type: boolean;
   is_errorable: boolean;
   name: MetadataTypeName;
} & BaseMetadataType;
export type MetadataFunctionArgumentDetailsDefinition =
   | null
   | ({ default_value: unknown } | { default_value: unknown; max_value: number; min_value: number });

export interface MetadataModuleBaseDefinition extends MetadataDefinition {
   version: string;
   versions?: { version: string }[];
   uuid: string;
}
export interface MetadataModuleDefinition extends MetadataModuleBaseDefinition {
   module_type: 'script';
   dependencies: MetadataModuleBaseDefinition[];
   objects: MetadataObjectDefinition[];
   constants: MetadataConstantDefinition[];
   enums: MetadataEnumDefinition[] | null;
   functions: MetadataFunctionDefinition[];
   classes: MetadataClassDefinition[];
   errors: MetadataErrorClassDefinition[];
   interfaces: MetadataInterfaceDefinition[];
   type_aliases: MetadataAliasDefinition[];
}
export interface MetadataObjectDefinition extends MetadataMemberDefinition, MetadataTypedDefinition {}
export interface MetadataConstantDefinition extends MetadataObjectDefinition {
   value?: unknown;
}
export interface MetadataEnumDefinition extends MetadataDefinition {
   constants: MetadataConstantDefinition[];
}
export interface Privilege {
   name: 'read_only' | 'none' | 'early_execution' | `${string}`;
}
export interface MetadataPropertyMemberDefinition extends MetadataMemberDefinition, MetadataTypedDefinition {
   get_privilege: Privilege[];
   set_privilege: Privilege[];
   is_baked: boolean;
}

export interface MetadataFunctionArgumentDefinition extends MetadataDefinition, MetadataTypedDefinition {
   details: MetadataFunctionArgumentDetailsDefinition | null;
}
export interface MetadataFunctionDefinition extends Omit<MetadataMemberDefinition, 'is_read_only'> {
   arguments: MetadataFunctionArgumentDefinition[];
   is_constructor: boolean;
   call_privilege: Privilege[];
   return_type: MetadataType;
}

export interface MetadataClassDefinition extends MetadataDefinition, MetadataTypedDefinition {
   base_types: MetadataType[];
   constants: MetadataConstantDefinition[];
   functions: MetadataFunctionDefinition[];
   properties: MetadataPropertyMemberDefinition[];
   iterator?: MetadataType;
}
export interface MetadataErrorClassDefinition extends MetadataDefinition, MetadataTypedDefinition {
   properties: MetadataPropertyMemberDefinition[];
}
export interface MetadataInterfaceDefinition extends MetadataDefinition, MetadataTypedDefinition {
   base_types: MetadataType[];
   properties: MetadataPropertyMemberDefinition[];
}

export type MetadataAliasMapEntry = { name: string; value: string };
export interface MetadataAliasDefinition extends MetadataDefinition, MetadataTypedDefinition {
   alias_type: 'type_map';
   mappings: MetadataAliasMapEntry[];
}